Lists now get updated correctyl on toggle.
1) Added a bunch of logic to handle the toggle states correctly.  The
issue was that ToggleButton.setChecked() calls the callback, which we
don't want.  We only want the callback called when the user selects the
button.
2)  Added lifecycle code to clean up the TaskListAdapter instances that
are attached to each TaskList and the TodayList.  Now the adapters will
be detached in onDestroy() or onDestroyView().
3) Added code to ensure that each list can only be attached to a single
adapter.  This is kind of limiting, but not for our purposes.
4) Created a ListProxy object so we can use proxies to attach to
ArrayAdapters, but I am not using it.  The idea would be to proxy an
underlying ArrayList, so we could filter or sort it and use the
filtered/sorted proxy for the ArrayAdapter.  This would allow us to use
a ArrayList in more than one ArrayAdapter ahd have it  have different
sorting while the the original array ordering is not changed.