Change the code to use id's rather than names
1) added Task.id() and TaskList.id() that return a string.  These
abstract the underlying datastore's id (ActiveAndroid) so we are less
tied to the underlying datastore.  For instance, ActiveAndroid using a
Long for the id, while parse using a String (for a GUID).
2) Changed various callbacks and systems that were storing names to use
ids.
This also gives us the flexibility to allow duplicate list or task names
and to move tasks between lists.